csa_app/pyproject.toml
[tool.poetry]
name = "control-self-assessment"
version = "2025.10.1"
description = "Self-assessment web application scaffold"
authors = ["Ferry Stelte <fstelte@stelte.nl>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
flask-sqlalchemy = "^3.1.1"
flask-login = "^0.6.3"
flask-wtf = "^1.2.1"
flask-migrate = "^4.0.5"
flask-talisman = "^1.1.0"
python-dotenv = "^1.0.1"
email-validator = "^2.1.0.post1"
pyotp = "^2.9.0"
psycopg = { extras = ["binary"], version = "^3.1.18" }

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-flask = "^1.3.0"
black = "^24.3.0"
ruff = "^0.6.8"
alembic = "^1.13.1"
pytest-cov = "^5.0.0"
bandit = "^1.7.7"
tox = "^4.12.1"

[build-system]
requires = ["poetry-core>=1.6.1"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "B", "UP"]
ignore = ["E203"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q"
testpaths = ["tests"]
